Refactor(DI): Implement dependency injection (#581)
* style: ran dotnet format

* refactor(core): implementing dependency injection

* feat(di): added dependency injection and dependency modules

* refactor(update): refactored update module

* refactor(di): migrated main, menu and sidebar to DI modules

* refactor(account): migrated account to DI module

* refactor(navigation): added navigation controller

* refactor(core): migrated MHR entities to async

* feat(core): migrated mhw to async

* feat(core): finished migrating core modules

* fix(di): fixed all DI modules and dependencies

* fix(backup): fixed route pointing to POST instead of PATCH

* fix(ui): fixed MainWindow not shutting down

* refactor(logging): organized and created new loggers

* refactor(localization): added new LocalizationRepository and deprecated old Localization singleton

* fix(di): fixed dependency registry returning empty arrays for unresolved singletons

* fix(core): fixed scanning and deadlocks

* feat(di): migrate feature flag repository to dependency injection

* style: run dotnet format

* refactor(core): migrated ScanManager to use new DI architecture

* fix(ui): fixed feature flag configuration not binding to UI changes

* fix(scan): fixed deadlock when running scanners while holding the read lock

* refactor(discord): migrated discord module to DI

* fix(core): stop scanner when game is closed

* refactor(DI): change module register log level to debug

+++ HunterPie.Integrations/Datasources/MonsterHunterRise/Definitions/MHRDualBladesStructure.cs
@@ -11,6 +11,6 @@ public struct MHRDualBladesStructure
     [FieldOffset(0x9EC)]
     [MarshalAs(UnmanagedType.Bool)] public bool IsDemonModeActive;
 
-    [FieldOffset(0xA38)] public long PiercingBindArrayPointer;
+    [FieldOffset(0xA38)] public nint PiercingBindArrayPointer;
     [FieldOffset(0xA48)] public bool IsArchDemonModeActive;
 }
