+++ HunterPie.Integrations/Datasources/MonsterHunterRise/Definitions/MHRPetalaceStructure.cs
@@ -5,10 +5,10 @@ namespace HunterPie.Integrations.Datasources.MonsterHunterRise.Definitions;
 [StructLayout(LayoutKind.Sequential)]
 public struct MHRPetalaceStructure
 {
-    public long Reference;
+    public nint Reference;
     public int Unk;
     public int Unk1;
     public int Unk2;
     public int Id;
-    public long Data;
+    public nint Data;
 }
